Extract messages from types
Messages are used to communicate between api and methods
